refactor(web): decompose comment-thread.tsx (816 -> 347) into focused siblings
The 802-line comment god-component is split the same way the artifact one was,
into siblings that own one concern each:

- comment-row.tsx   — CommentRow: one comment (body / inline edit / reactions /
                      the react·edit·copy·delete hover toolbar)
- comment-composer.tsx — MentionField (the @mention autocomplete input) + Composer
                      (the new-comment box that wraps it)
- comment-thread.tsx — keeps the thread/layout trio that belongs together:
                      PinnedZone (margin layout), CommentCard (a thread card),
                      ResolvedSection

Pure code movement, no behaviour change. comment-thread re-exports Composer so
comment-panels' import path is untouched. Typecheck clean, the 24 web unit tests
still pass, biome + knip clean.
